Dont fail builds on failure to prefetch query
