+++ graphql-codegen/graphql-plugin-prefetch.ts
@@ -10,59 +10,13 @@ type FileQueries = {
 };
 
 const template = (imports: string[], functions: string[]) => {
-	return `
-import { QueryClient } from '@tanstack/react-query';
-import makeQueryClient from '~lib/makeQueryClient';
-
-${imports.join('\n')}
+	return `${imports.join('\n')}
 
 type Fn<T> = (vars: T) => Promise<unknown>;
-type Key = keyof typeof fns;
-type Vars = {
-	[K in Key]?: Parameters<typeof fns[K]>[0];
-};
 
-const options = {
-    cacheTime: 24 * 60 * 60 * 1000,
-};
-
-async function doPrefetch<T extends Key>(k: T, v: Vars[T], client: QueryClient) {
-	const r = await fns[k](v as any);
-
-	const hasVars = Object.keys(v as any).length > 0;
-	const queryKey = hasVars ? [k, v] : [k];
-	const infiniteQueryKey = hasVars ? [\`\${k}.infinite\`, v] : [\`\${k}.infinite\`];
-
-    await client.prefetchQuery({
-        queryKey,
-        queryFn: () => r,
-        ...options
-    });
-    await client.prefetchInfiniteQuery({
-        queryKey: infiniteQueryKey,
-        queryFn: () => r,
-        initialPageParam: null,
-        ...options
-    });
-}
-
-export async function prefetchQueries(
-	vars: Vars,
-	client: QueryClient = makeQueryClient(),
-): Promise<QueryClient> {
-	const queries = Object.keys(vars) as Key[];
-
-	await Promise.all(
-		queries.map(k => doPrefetch(k, vars[k], client))
-	);
-	
-	return client;
-}
-
-const fns = {
+export const fns = {
 	${functions.join('\n\t')}
-};
-`;
+};`;
 };
 
 function capitalize(string: string) {
