+++ src/pages/[language]/discover/collections.ts
@@ -15,7 +15,7 @@ import getIntl from '~lib/getIntl';
 import { getLanguageIdByRoute } from '~lib/getLanguageIdByRoute';
 import { getLanguageRoutes } from '~lib/getLanguageRoutes';
 import root from '~lib/routes';
-import { prefetchQueries } from '~src/__generated__/prefetch';
+import { prefetchQueries } from '~src/lib/prefetchQueries';
 import serializableDehydrate from '~src/lib/serializableDehydrate';
 
 export default DiscoverCollections;
