+++ src/pages/[language]/library/playlists/[playlist]/index.ts
@@ -3,8 +3,8 @@ import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 
 import LibraryPlaylistDetail from '~containers/library/playlist/libraryPlaylistDetail';
 import { storeRequest } from '~lib/api/storeRequest';
-import { prefetchQueries } from '~src/__generated__/prefetch';
 import getDehydratedProps from '~src/lib/getDehydratedProps';
+import { prefetchQueries } from '~src/lib/prefetchQueries';
 
 export default LibraryPlaylistDetail;
